+++ linear_model/__init__.py
@@ -1,3 +1,0 @@
-from ._LinearRegression import LinearRegression
-from ._LogisticRegression import LogisticRegression
-__all__ = ["LinearRegression", "LogisticRegression"]