+++ sklearn_clone/linear_model/__init__.py
@@ -1,0 +1,4 @@
+from ._LinearRegression import LinearRegression
+from ._LogisticRegression import LogisticRegression
+from ._regulizations import Ridge, Lasso, ElasticNet
+__all__ = ["LinearRegression", "LogisticRegression", "Ridge", "Lasso", "ElasticNet"]